Refactored Method and Function classes to inherit from a Callable superclass instead of Function inheriting from Method.
