+++ SharpGen/Generator/DefaultGenerators.cs
@@ -41,10 +41,10 @@ namespace SharpGen.Generator
         public IMultiCodeGenerator<CsField, MemberDeclarationSyntax> ExplicitOffsetField { get; }
         public IMultiCodeGenerator<CsField, MemberDeclarationSyntax> AutoLayoutField { get; }
         public IMultiCodeGenerator<CsStruct, MemberDeclarationSyntax> Struct { get; }
-        public ICodeGenerator<CsMethod, ExpressionSyntax> NativeInvocation { get; }
+        public ICodeGenerator<CsCallable, ExpressionSyntax> NativeInvocation { get; }
         public IMultiCodeGenerator<CsParameter, StatementSyntax> ParameterProlog { get; }
         public IMultiCodeGenerator<CsParameter, StatementSyntax> ParameterEpilog { get; }
-        public IMultiCodeGenerator<CsMethod, MemberDeclarationSyntax> Callable { get; }
+        public IMultiCodeGenerator<CsCallable, MemberDeclarationSyntax> Callable { get; }
         public IMultiCodeGenerator<CsMethod, MemberDeclarationSyntax> Method { get; }
         public IMultiCodeGenerator<CsFunction, MemberDeclarationSyntax> Function { get; }
         public IMultiCodeGenerator<CsInterface, MemberDeclarationSyntax> Interface { get; }
